rename fetcher -> dagfetcher, ordering -> dagordering
