+++ gossip/dagordering/ordering_test.go
@@ -1,4 +1,4 @@
-package ordering
+package dagordering
 
 import (
 	"errors"
